Update main.py
Added Login/Signup for the user. It checks against our own database to see if a similar email exists. If it does: It asks the user to Login or use a different email. During Signup user needs to fill the following info:
Email
Pw
First Name
Last Name
Phone Number